requirements.txt
numpy
pandas
typing-extensions
python-dotenv
mysql-connector-python
pymysql
scikit-learn
# -e . 